updating everything but cumulusci.yml
